+++ build-logic/src/main/java/com/squareup/workflow1/buildsrc/internal/catalogs.kt
@@ -1,4 +1,4 @@
-package com.squareup.workflow1
+package com.squareup.workflow1.buildsrc.internal
 
 import org.gradle.api.Project
 import org.gradle.api.artifacts.MinimalExternalModuleDependency
