--- build-logic/src/main/java/compose-ui-tests.gradle.kts
@@ -1,12 +1,0 @@
-import com.squareup.workflow1.library
-import com.squareup.workflow1.libsCatalog
-
-plugins {
-  id("android-defaults")
-}
-
-dependencies {
-  "androidTestImplementation"(project(":workflow-ui:internal-testing-compose"))
-
-  "androidTestImplementation"(libsCatalog.library("androidx-compose-ui-test-junit4"))
-}
